requirements.txt

pandas
pillow